+++ server/src/main/java/org/allaymc/server/block/component/BlockBeaconBaseComponentImpl.java
@@ -11,21 +11,25 @@ import org.allaymc.api.world.Dimension;
 /**
  * @author daoge_cmd
  */
-public class BlockCraftingTableBaseComponentImpl extends BlockBaseComponentImpl {
-    public BlockCraftingTableBaseComponentImpl(BlockType<? extends BlockBehavior> blockType) {
+public class BlockBeaconBaseComponentImpl extends BlockBaseComponentImpl {
+    public BlockBeaconBaseComponentImpl(BlockType<? extends BlockBehavior> blockType) {
         super(blockType);
     }
 
     @Override
     public boolean onInteract(ItemStack itemStack, Dimension dimension, PlayerInteractInfo interactInfo) {
-        if (super.onInteract(itemStack, dimension, interactInfo)) return true;
+        if (super.onInteract(itemStack, dimension, interactInfo)) {
+            return true;
+        }
 
         var player = interactInfo.player();
-        if (player.isSneaking()) return false;
-
-        var craftingTableContainer = player.getContainer(FullContainerType.CRAFTING_TABLE);
-        craftingTableContainer.setBlockPos(new Position3i(interactInfo.clickedBlockPos(), interactInfo.player().getDimension()));
-        craftingTableContainer.addViewer(player);
+        if (player.isSneaking()) {
+            return false;
+        }
+        
+        var beaconContainer = player.getContainer(FullContainerType.BEACON);
+        beaconContainer.setBlockPos(new Position3i(interactInfo.clickedBlockPos(), interactInfo.player().getDimension()));
+        beaconContainer.addViewer(player);
         return true;
     }
 }