+++ server/src/main/java/org/allaymc/server/block/component/BlockEnchantingTableBaseComponentImpl.java
@@ -5,6 +5,7 @@ import org.allaymc.api.block.dto.PlayerInteractInfo;
 import org.allaymc.api.block.type.BlockType;
 import org.allaymc.api.container.FullContainerType;
 import org.allaymc.api.item.ItemStack;
+import org.allaymc.api.math.position.Position3i;
 import org.allaymc.api.world.Dimension;
 
 /**
@@ -20,8 +21,12 @@ public class BlockEnchantingTableBaseComponentImpl extends BlockBaseComponentImp
         if (super.onInteract(itemStack, dimension, interactInfo)) return true;
 
         var player = interactInfo.player();
+        if (player.isSneaking()) {
+            return false;
+        }
+
         var enchantTableContainer = player.getContainer(FullContainerType.ENCHANT_TABLE);
-        enchantTableContainer.setBlockPos(interactInfo.clickedBlockPos());
+        enchantTableContainer.setBlockPos(new Position3i(interactInfo.clickedBlockPos(), interactInfo.player().getDimension()));
         enchantTableContainer.addViewer(player);
         return true;
     }