+++ server/src/main/java/org/allaymc/server/blockentity/impl/BlockEntityBeaconImpl.java
@@ -1,0 +1,25 @@
+package org.allaymc.server.blockentity.impl;
+
+import lombok.experimental.Delegate;
+import org.allaymc.api.blockentity.component.BlockEntityBeaconBaseComponent;
+import org.allaymc.api.blockentity.initinfo.BlockEntityInitInfo;
+import org.allaymc.api.blockentity.interfaces.BlockEntityBeacon;
+import org.allaymc.api.component.interfaces.Component;
+import org.allaymc.server.component.interfaces.ComponentProvider;
+
+import java.util.List;
+
+/**
+ * @author daoge_cmd
+ */
+public class BlockEntityBeaconImpl extends BlockEntityImpl implements BlockEntityBeacon {
+    public BlockEntityBeaconImpl(BlockEntityInitInfo initInfo, List<ComponentProvider<? extends Component>> componentProviders) {
+        super(initInfo, componentProviders);
+    }
+
+    @Delegate
+    @Override
+    public BlockEntityBeaconBaseComponent getBaseComponent() {
+        return (BlockEntityBeaconBaseComponent) super.getBaseComponent();
+    }
+}